Ground reward redemptions in durable accounting
The point shop needs DB-managed items, whole-point balances, and a durable audit trail before purchase, admin, or notification work can be split into later PRs. This commit replaces the static catalog with reward shop persistence, seeds the coffee voucher, introduces member roles for the future admin surface, and records point grants as immutable history entries.

Constraint: This is PR 1 of the split implementation plan and must avoid user-facing redemption flows until the accounting foundation is merged.

Constraint: Item availability uses active plus stock and per-member purchase usage; no separate sellable column is introduced.

Rejected: Keep RewardShopCatalog static data | would make inventory and purchase limits impossible to operate from DB state.

Rejected: Track both sold_quantity and consumed_quantity | duplicate counters would blur the source of truth for stock restoration.

Confidence: high

Scope-risk: moderate

Directive: Preserve sold_quantity as the item-level stock counter and purchase_count as the member-level limit counter when adding redemption transitions.

Tested: ./gradlew test --tests '*Reward*' --tests '*Member*' --console=plain

Tested: ./gradlew check --console=plain

Not-tested: Browser/UI flows and outbound email notifications; those belong to later PRs in the train.

+++ backend/src/main/java/coin/coinzzickmock/feature/account/application/result/AccountSummaryResult.java
@@ -8,6 +8,6 @@ public record AccountSummaryResult(
         double available,
         double totalUnrealizedPnl,
         double roi,
-        double rewardPoint
+        int rewardPoint
 ) {
 }
